big refactoring in progress
