+++ docs/changelog.rst
@@ -15,7 +15,7 @@ Version 3.0
     - Tecnai Polara (WinXP, Python 3.4)
     - Glacios (Win10, Python 3.8)
     - Tundra (Win10, Python 3.11)
-    - Titan Krios G1 (Win7, Python 3.6), G2, G3i (Win10, Python 3.8)
+    - Titan Krios G1 (Win7, Python 3.6), G2, G3i (Win10, Python 3.8), G4 (Win10, Python 3.8)
 
 Version 2.0.0
 ^^^^^^^^^^^^^
