+++ docs/microscope.rst
@@ -6,35 +6,23 @@ Microscope class
 The :class:`Microscope` class provides a Python interface to the microscope.
 Below are the main class properties, each represented by a separate class:
 
-    * acquisition = :meth:`~pytemscript.microscope.Acquisition`
-    * apertures = :meth:`~pytemscript.microscope.Apertures`
-    * autoloader = :meth:`~pytemscript.microscope.Autoloader`
-    * detectors = :meth:`~pytemscript.microscope.Detectors`
-    * energy_filter = :meth:`~pytemscript.microscope.EnergyFilter`
-    * gun = :meth:`~pytemscript.microscope.Gun`
-    * optics = :meth:`~pytemscript.microscope.Optics`
+    * acquisition = :meth:`~pytemscript.modules.Acquisition`
+    * apertures = :meth:`~pytemscript.modules.Apertures`
+    * autoloader = :meth:`~pytemscript.modules.Autoloader`
+    * detectors = :meth:`~pytemscript.modules.Detectors`
+    * energy_filter = :meth:`~pytemscript.modules.EnergyFilter`
+    * gun = :meth:`~pytemscript.modules.Gun`
+    * optics = :meth:`~pytemscript.modules.Optics`
 
-        * illumination = :meth:`~pytemscript.microscope.Illumination`
-        * projection = :meth:`~pytemscript.microscope.Projection`
+        * illumination = :meth:`~pytemscript.modules.Illumination`
+        * projection = :meth:`~pytemscript.modules.Projection`
 
-    * piezo_stage = :meth:`~pytemscript.microscope.PiezoStage`
-    * stage = :meth:`~pytemscript.microscope.Stage`
-    * stem = :meth:`~pytemscript.microscope.Stem`
-    * temperature = :meth:`~pytemscript.microscope.Temperature`
-    * user_door = :meth:`~pytemscript.microscope.UserDoor`
-    * vacuum = :meth:`~pytemscript.microscope.Vacuum`
-
-
-Image object
-------------
-
-Two acquisition functions: :meth:`~pytemscript.microscope.Acquisition.acquire_tem_image` and
-:meth:`~pytemscript.microscope.Acquisition.acquire_stem_image` return an :class:`Image` object
-that has the following methods and properties:
-
-.. autoclass:: pytemscript.microscope.Image
-    :members: width, height, bit_depth, pixel_type, data, save, name, metadata
-
+    * piezo_stage = :meth:`~pytemscript.modules.PiezoStage`
+    * stage = :meth:`~pytemscript.modules.Stage`
+    * stem = :meth:`~pytemscript.modules.Stem`
+    * temperature = :meth:`~pytemscript.modules.Temperature`
+    * user_door = :meth:`~pytemscript.modules.UserDoor`
+    * vacuum = :meth:`~pytemscript.modules.Vacuum`
 
 Example usage
 -------------
@@ -51,9 +39,19 @@ Example usage
     defocus = microscope.optics.projection.defocus
     microscope.optics.normalize_all()
 
+
+Image object
+------------
+
+Two acquisition functions: :meth:`~pytemscript.modules.Acquisition.acquire_tem_image` and
+:meth:`~pytemscript.modules.Acquisition.acquire_stem_image` return an :class:`Image` object
+that has the following methods and properties:
+
+.. autoclass:: pytemscript.modules.Image
+    :members: width, height, bit_depth, pixel_type, data, save, name, metadata
+
 Documentation
 -------------
 
-.. automodule:: pytemscript.microscope
-    :members:
-    :exclude-members: Image
+.. automodule:: pytemscript.modules
+    :members: Acquisition, Apertures, Autoloader, Detectors, EnergyFilter, Gun, Optics, Illumination, Projection, PiezoStage, Stage, Stem, Temperature, UserDoor, Vacuum
