+++ pytemscript/modules/autoloader.py
@@ -1,0 +1,120 @@
+from ..utils.enums import CassetteSlotStatus
+
+
+class Autoloader:
+    """ Sample loading functions. """
+    def __init__(self, client):
+        self._client = client
+        self._has_autoloader_adv = None
+        self._err_msg = "Autoloader is not available"
+        self._err_msg_adv = "This function is not available in your advanced scripting interface."
+
+    @property
+    def __adv_available(self):
+        if self._has_autoloader_adv is None:
+            self._has_autoloader_adv = self._client.has("tem_adv.AutoLoader")
+        return self._has_autoloader_adv
+
+    @property
+    def is_available(self):
+        """ Status of the autoloader. Should be always False on Tecnai instruments. """
+        return self._client.get("tem.AutoLoader.AutoLoaderAvailable")
+
+    @property
+    def number_of_slots(self):
+        """ The number of slots in a cassette. """
+        if self.is_available:
+            return self._client.get("tem.AutoLoader.NumberOfCassetteSlots")
+        else:
+            raise RuntimeError(self._err_msg)
+
+    def load_cartridge(self, slot):
+        """ Loads the cartridge in the given slot into the microscope.
+
+        :param slot: Slot number
+        :type slot: int
+        """
+        if self.is_available:
+            total = self.number_of_slots
+            slot = int(slot)
+            if slot > total:
+                raise ValueError("Only %s slots are available" % total)
+            if self.slot_status(slot) != CassetteSlotStatus.OCCUPIED.name:
+                raise RuntimeError("Slot %d is not occupied" % slot)
+            self._client.call("tem.AutoLoader.LoadCartridge()", slot)
+        else:
+            raise RuntimeError(self._err_msg)
+
+    def unload_cartridge(self):
+        """ Unloads the cartridge currently in the microscope and puts it back into its
+        slot in the cassette.
+        """
+        if self.is_available:
+            self._client.call("tem.AutoLoader.UnloadCartridge()")
+        else:
+            raise RuntimeError(self._err_msg)
+
+    def run_inventory(self):
+        """ Performs an inventory of the cassette.
+        Note: This function takes considerable time to callute.
+        """
+        # TODO: check if cassette is present
+        if self.is_available:
+            self._client.call("tem.AutoLoader.PerformCassetteInventory()")
+        else:
+            raise RuntimeError(self._err_msg)
+
+    def slot_status(self, slot):
+        """ The status of the slot specified.
+
+        :param slot: Slot number
+        :type slot: int
+        """
+        if self.is_available:
+            total = self.number_of_slots
+            if slot > total:
+                raise ValueError("Only %s slots are available" % total)
+            status = self._client.call("tem.AutoLoader.SlotStatus()", int(slot))
+            return CassetteSlotStatus(status).name
+        else:
+            raise RuntimeError(self._err_msg)
+
+    def undock_cassette(self):
+        """ Moves the cassette from the docker to the capsule. """
+        if self.__adv_available:
+            if self.is_available:
+                self._client.call("tem_adv.AutoLoader.UndockCassette()")
+            else:
+                raise RuntimeError(self._err_msg)
+        else:
+            raise NotImplementedError(self._err_msg_adv)
+
+    def dock_cassette(self):
+        """ Moves the cassette from the capsule to the docker. """
+        if self.__adv_available:
+            if self.is_available:
+                self._client.call("tem_adv.AutoLoader.DockCassette()")
+            else:
+                raise RuntimeError(self._err_msg)
+        else:
+            raise NotImplementedError(self._err_msg_adv)
+
+    def initialize(self):
+        """ Initializes / Recovers the Autoloader for further use. """
+        if self.__adv_available:
+            if self.is_available:
+                self._client.call("tem_adv.AutoLoader.Initialize()")
+            else:
+                raise RuntimeError(self._err_msg)
+        else:
+            raise NotImplementedError(self._err_msg_adv)
+
+    def buffer_cycle(self):
+        """ Synchronously runs the Autoloader buffer cycle. """
+        if self.__adv_available:
+            if self.is_available:
+                self._client.call("tem_adv.AutoLoader.BufferCycle()")
+            else:
+                raise RuntimeError(self._err_msg)
+        else:
+            raise NotImplementedError(self._err_msg_adv)
