+++ pytemscript/utils/misc.py
@@ -6,6 +6,7 @@ import zlib
 import gzip
 import io
 import functools
+import logging
 
 
 MIME_TYPE_PICKLE = "application/python-pickle"
@@ -44,7 +45,7 @@ ARRAY_ENDIANNESS = {"LITTLE", "BIG"}
 
 def unpack_array(obj):
     """
-    Unpack an packed array.
+    Unpack a packed array.
 
     :param obj: Dict with packed array
     """
@@ -108,45 +109,20 @@ def gzip_decode(content):
     return zlib.decompress(content, 16 + zlib.MAX_WBITS)    # No keyword arguments until Python 3.6
 
 
-def multi_getattr(obj, attr):
-    attributes = attr.split(".")
-    for i in attributes:
-        try:
-            obj = getattr(obj, i)
-            if callable(obj):
-                obj = obj()
-        except AttributeError:
-            raise
-    return obj
-
+def rgetattr(obj, attr, *args, **kwargs):
+    """ Recursive getattr or callable on a COM object"""
+    try:
+        if kwargs.get("typeget", True):
+            logging.info("Getattr: %s, args=%s, kwargs=%s" %
+                         (attr, args, kwargs))
+        result = functools.reduce(getattr, attr.split('.'), obj)
+        return result(*args, **kwargs) if callable(result) else result
+    except AttributeError:
+        logging.error("Attribute error %s" % attr)
+        raise AttributeError("AttributeError: %s" % attr)
 
 def rsetattr(obj, attr, val):
     """ https://stackoverflow.com/a/31174427 """
+    logging.info("Setattr: %s=%s" % (attr, val))
     pre, _, post = attr.rpartition('.')
-    return setattr(rgetattr(obj, pre) if pre else obj, post, val)
-
-
-def rgetattr(obj, attr):
-    def _getattr(obj, attr):
-        return getattr(obj, attr)
-    result = functools.reduce(_getattr, [obj] + attr.split('.'))
-    return result
-
-
-def rexecattr(obj, attr, *args, **kwargs):
-    def _getattr(obj, attr):
-        return getattr(obj, attr)
-    result = functools.reduce(_getattr, [obj] + attr.split('.'))
-    if args or kwargs:
-        return result(*args, **kwargs)
-    else:
-        return result()
-
-
-def rhasattr(obj, attr):
-    """ https://stackoverflow.com/a/65781864 """
-    try:
-        functools.reduce(getattr, attr.split("."), obj)
-        return True
-    except AttributeError:
-        return False
+    return setattr(rgetattr(obj, pre, typeget=False) if pre else obj, post, val)
